L10N follow up
This pull request is a follow up from the first one that was merged.
In this pull request I have done some clean up on the translation file
Also, some of which we no longer use I have remove them from the file as
well.

I also did update the i18n module to the latest version where this
version include unit test and some great updates.

The polymer template still not possible to localize with the current
tools that we have for Webmaker, but I did have some idea and planning
in which will require some work to make it possible and ideal in our
case.

You can check the user-state.html file which is not an ideal way of
doing it at all and that is not going to work (Only works on the first
page load but if we did change the state again might break).

I will have another follow up Pull request or will hack on it together
at the workweek.